perf: optimize memory usage and resource consumption
Backend:
- Use Multer diskStorage instead of memory for file uploads (Orders, Products, Rewards)
- Add multer util helpers for buffer read and temp file cleanup
- Add default limit (100) to findPublicRewards query
- Add XLSX streaming note for future optimization

Frontend:
- Lazy-load recharts via dynamic import in analytics page
- Lazy-load qrcode in setup page (dynamic import on QR generation)
- Lazy-load xlsx in ImportSpecialPricesModal
- Lazy-load @hello-pangea/dnd in categorias via CategoryTreeWithDnd
- Reduce React Query gcTime from 10 to 5 minutes
- Add maxPages: 5 to useProductsInfinite
- Limit CartContext migrateCartItems concurrency to 3 requests

Made-with: Cursor
