+++ apps/backend/src/modules/products/products.module.ts
@@ -4,14 +4,14 @@ import { ProductsService } from './products.service';
 import { ProductsController } from './products.controller';
 import { PrismaModule } from '../prisma/prisma.module';
 import { DatabaseModule } from '../database/database.module';
-// import { CacheModule } from '@nestjs/cache-manager';
+import { multerDiskStorage } from '../../common/multer-disk.config';
 
 @Module({
   imports: [
     PrismaModule,
     DatabaseModule,
     MulterModule.register({
-      storage: require('multer').memoryStorage(),
+      storage: multerDiskStorage,
       limits: {
         fileSize: 10 * 1024 * 1024, // 10MB
       },
