+++ apps/frontend/src/app/providers.tsx
@@ -10,7 +10,7 @@ export function Providers({ children }: { children: React.ReactNode }) {
     defaultOptions: {
       queries: {
         staleTime: 5 * 60 * 1000, // 5 minutos
-        gcTime: 10 * 60 * 1000, // 10 minutos (anteriormente cacheTime)
+        gcTime: 5 * 60 * 1000, // 5 minutos (reduce retención en memoria)
         retry: 1,
         refetchOnWindowFocus: false,
       },
